requirements.txt
numpy==2.0.0
python-dotenv==1.0.1
google-generativeai==0.8.3